Speed settings overhaul
Split the single movement speed setting into multiple settings to
improve ease of use and customizability. Base speed, speed limit, and
mount speed are all now directly set instead of an offset to retail
values.

Animation speed setting added that controls player movement animation
speed. This is a little hard to explain. Take your current speed and
divide it by your base speed. Normally this would be 50/50=1.0, this is
the rate at which your running animation plays, 1x. Flee brings you to
80/50=1.6x animation speed. If you increase base speed to 80, then you
still have a 1x animation speed. This setting is applied as a divisor to
the base speed, so if you want to increase the base movement speed and
also increase animation speed, you would increase this value (e.g. to
make it always appear as if the player has a flee effect you would set
BASE_SPEED to 80 and ANIMATION_SPEED_DIVISOR to 1.6).

Speedsub was changed to animationSpeed to reflect this, and baseSpeed
was added to track the base speed of an entity. For most entities these
will be the same, but allows for player animation adjustments.